big refactor (part 6)

+++ app/controllers/users_controller.rb
@@ -92,8 +92,12 @@ class UsersController < ApplicationController
     if current_service.user_id == @user.id
       flash[:error] = "You can't remove association because the user is the owner"
     else
-      if UserAssociation.where(user: @user, associable: current_service).destroy_all
-        flash[:success] = I18n.t('actions.success.destroyed', resource: t('activerecord.models.user_association'))
+      if [current_service.main_admin, current_service.main_developer, current_service.main_commercial, current_service.main_accountant].include?(@user)
+        flash[:error] = "You can't remove association because the user is an official"
+      else
+        if UserAssociation.where(user: @user, associable: current_service).destroy_all
+          flash[:success] = I18n.t('actions.success.destroyed', resource: t('activerecord.models.user_association'))
+        end
       end
     end
     redirect_back(fallback_location: user_path(@user))
